--- include/nxtgm/optimizers/gm/discrete/graph_cut.hpp
@@ -1,67 +1,0 @@
-#pragma once
-
-#include <chrono>
-#include <nxtgm/optimizers/gm/discrete/optimizer_base.hpp>
-
-#include <nxtgm/plugins/min_st_cut/min_st_cut_base.hpp>
-
-namespace nxtgm
-{
-
-class GraphCut : public DiscreteGmOptimizerBase
-{
-    class parameters_type : public OptimizerParametersBase
-    {
-      public:
-        inline parameters_type(const nlohmann::json &json_parameters)
-            : OptimizerParametersBase(json_parameters)
-        {
-            if (json_parameters.contains("min_st_cut_plugin_name"))
-            {
-                min_st_cut_plugin_name = json_parameters["min_st_cut_plugin_name"].get<std::string>();
-            }
-            if (json_parameters.contains("submodular_epsilon"))
-            {
-                submodular_epsilon = json_parameters["submodular_epsilon"].get<double>();
-            }
-        }
-
-        std::string min_st_cut_plugin_name;
-        double submodular_epsilon = 1e-6;
-    };
-
-  public:
-    using base_type = DiscreteGmOptimizerBase;
-    using solution_type = typename DiscreteGm::solution_type;
-
-    using reporter_callback_wrapper_type = typename base_type::reporter_callback_wrapper_type;
-    using repair_callback_wrapper_type = typename base_type::repair_callback_wrapper_type;
-
-    using base_type::optimize;
-
-    inline static std::string name()
-    {
-        return "GraphCut";
-    }
-    virtual ~GraphCut() = default;
-
-    GraphCut(const DiscreteGm &gm, const nlohmann::json &parameters);
-
-    OptimizationStatus optimize(reporter_callback_wrapper_type &, repair_callback_wrapper_type &,
-                                const_discrete_solution_span starting_point) override;
-
-    SolutionValue best_solution_value() const override;
-    SolutionValue current_solution_value() const override;
-
-    const solution_type &best_solution() const override;
-    const solution_type &current_solution() const override;
-
-  private:
-    parameters_type parameters_;
-
-    SolutionValue best_solution_value_;
-    solution_type best_solution_;
-
-    std::unique_ptr<MinStCutBase> min_st_cut_;
-};
-} // namespace nxtgm
