+++ include/nxtgm/optimizers/gm/discrete/optimizer_base.hpp
@@ -1,9 +1,16 @@
 #pragma once
 
 #include <nxtgm/models/gm/discrete_gm.hpp>
+#include <nxtgm/optimizers/callback_base.hpp>
 #include <nxtgm/optimizers/optimizer_base.hpp>
 
 #include <nlohmann/json.hpp>
+#include <xplugin/xfactory.hpp>
+
+#include <iostream>
+#include <memory>
+#include <string>
+#include <unordered_map>
 
 namespace nxtgm
 {
@@ -23,13 +30,32 @@ class DiscreteGmOptimizerBase : public OptimizerBase<DiscreteGm, DiscreteGmOptim
     }
 };
 
-class DiscreteOptimizerFactoryBase
+class DiscreteGmOptimizerFactoryBase
 {
   public:
-    virtual ~DiscreteOptimizerFactoryBase() = default;
+    virtual ~DiscreteGmOptimizerFactoryBase() = default;
 
-    virtual std::unique_ptr<DiscreteGmOptimizerBase> create(const DiscreteGm &gm) const = 0;
-    virtual std::unique_ptr<DiscreteOptimizerFactoryBase> clone() const = 0;
+    virtual std::unique_ptr<DiscreteGmOptimizerBase> create(const DiscreteGm &gm,
+                                                            const nlohmann::json &params) const = 0;
+
+    static std::string plugin_type()
+    {
+        return "discrete_gm_optimizer";
+    }
+
+    static std::string plugin_dir_env_var()
+    {
+        return "NXTGM_DISCRETE_GM_OPTIMIZER_PLUGIN_PATH";
+    }
+
+    // priority of the plugin (higher means more important)
+    virtual int priority() const = 0;
+
+    // license of the plugin
+    virtual std::string license() const = 0;
+
+    // description of the plugin
+    virtual std::string description() const = 0;
 };
 
 } // namespace nxtgm
