+++ src/nxtgm/optimizers/gm/discrete/optimizer_base.cpp
@@ -1,0 +1,6 @@
+#include <nxtgm/optimizers/gm/discrete/optimizer_base.hpp>
+
+namespace nxtgm
+{
+
+}
