+++ src/nxtgm/plugins/CMakeLists.txt
@@ -2,3 +2,4 @@ add_custom_target(all_plugins)
 
 add_subdirectory(min_st_cut)
 add_subdirectory(qpbo)
+add_subdirectory(discrete_gm_optimizer)
