+++ tests/optimizers/gm/discrete/test_matching_icm.cpp
@@ -1,5 +1,4 @@
 #include <discrete_gm_optimizer_tester.hpp>
-#include <nxtgm/optimizers/gm/discrete/matching_icm.hpp>
 
 TEST_CASE("matching-icm")
 {
